test(client): add remaining functions

+++ client/src/main/java/dev/caceresenzo/gitbook/model/ChangeRequest.java
@@ -16,7 +16,7 @@ public class ChangeRequest {
 	private String subject;
 
 	@JsonProperty("number")
-	private long number;
+	private int number;
 
 	@JsonProperty("status")
 	private Status status;
